pyproject.toml
[tool.poetry]
name = "antigravirt"
version = "0.1.0"
description = "Universal Data Analysis Agent"
authors = ["Antigravirt Team"]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
websockets = "^12.0"
langchain = "^0.3.0"
langchain-core = "^0.3.0"
langchain-community = "^0.3.0"
langchain-google-genai = "^2.0.0"
langchain-openai = "^0.2.0"
langgraph = "^0.2.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
mcp = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I"]
ignore = []
